updated run_bot.py moved files to /run folder
