+++ .gitignore
@@ -1,2 +1,3 @@
 work
 __pycache__
+config.json
