--- main.py
@@ -1,4 +1,0 @@
-from run_bot import run_bot
-
-if __name__ == "__main__":
-    run_bot(batch_size=3)
