+++ run/main.py
@@ -1,2 +1,4 @@
+from run_bot import run_bot
+
 if __name__ == "__main__":
-    pass
+    run_bot(batch_size=3)
